fix: type 변경을 했음에도 DTO의 타입은 그대로여서 api 호출 시 에러가 난 버그 해결 #21

+++ src/main/java/com/OOL/oolfinance/dto/main/StockTableDTO.java
@@ -36,6 +36,6 @@ public class StockTableDTO {
     @Schema(name = "tradingValue", type = "BigDecimal", description = "거래대금")
     private BigDecimal tradingValue;
 
-    @Schema(name = "tradingVolume", type = "long", description = "거래량")
-    private long tradingVolume;
+    @Schema(name = "tradingVolume", type = "BigDecimal", description = "거래량")
+    private BigDecimal tradingVolume;
 }
